+++ components/article/ArticleList.tsx
@@ -1,48 +1,14 @@
 'use client';
 import { getArticlesAPI } from '@/services/articles';
 import ArticlePreview from './ArticlePreview';
-import { useInfiniteQuery } from '@tanstack/react-query';
-import React, { useEffect, useRef } from 'react';
-import { flexCenter, greenButton } from '@/styles/common.css';
-import useIntersectionObserver from '@/hooks/useIntersectionObserver';
+import React, { useRef } from 'react';
+import { flexCenter } from '@/styles/common.css';
+import useInfiniteScroll from '@/hooks/useInfiniteScroll';
 
 const ArticleList = () => {
   const targetRef = useRef(null);
 
-  const { data, error, fetchNextPage, hasNextPage, isFetching, isFetchingNextPage, status } = useInfiniteQuery({
-    queryKey: ['articles'],
-    queryFn: ({ pageParam = 1 }) => {
-      console.log(pageParam);
-
-      return getArticlesAPI(pageParam);
-    },
-    getNextPageParam: (lastPage, pages) => {
-      console.log('getNextPageParam 발동');
-
-      const totalPage = Math.ceil(lastPage.articlesCount / 10);
-      let currentPage = pages.length;
-      if (totalPage < pages.length) {
-        console.log('그만');
-
-        return undefined;
-      }
-      // console.log(totalPage);
-      // console.log(currentPage);
-
-      return currentPage++;
-    },
-    retry: false,
-    refetchOnWindowFocus: false,
-  });
-
-  const nextPage = () => {
-    if (!hasNextPage || isFetchingNextPage) {
-      return;
-    }
-    fetchNextPage();
-  };
-
-  useIntersectionObserver(nextPage, targetRef);
+  const { data } = useInfiniteScroll(getArticlesAPI, targetRef);
 
   return (
     <div>
